:recycle: (refactor) move next block factory to own folder
