+++ libs/functions/chatbot/src/lib/services/next-block/next-block.factory.ts
