+++ libs/functions/chatbot/src/lib/services/next-block/next-block.interface.ts
